==> code/04b_peaks_plot_metrics.ipynb <==
# Analysis of distribution of shape statistics

**Part 2: plot metrics**

This notebook visualizes the numerical results of the peak-finding procedure from the previous notebook, `04a_peaks_compute`. 

import os

os.environ["USE_PYGEOS"] = "0"
import json
import geopandas
import pandas
import numpy
import matplotlib.pyplot as plt
from matplotlib import rcParams
import seaborn as sns
from palettable.cartocolors.qualitative import Bold_6

from scipy.signal import find_peaks
from scipy.stats import gaussian_kde

Set default plotting theme and add color dictionary for continents.

sns.set_theme(
    context="paper",
    style="ticks",
    rc={
        "patch.force_edgecolor": False,
        "axes.spines.top": False,
        "axes.spines.right": False,
        "axes.grid": True,
    },
    palette=Bold_6.hex_colors,
)

coldict = {
    "Africa": 0,
    "Asia": 1,
    "Europe": 2,
    "North America": 3,
    "Oceania": 4,
    "South America": 5,
}

Load results from previous notebook

results = json.load(open("../results/04_peaks_results.json"))

# load parameters (metric names)
options = [
    "circular_compactness_index",
    "isoperimetric_quotient_index",
    "isoareal_quotient_index",
    "radii_ratio_index",
    "diameter_ratio_index",
]

Load the cities / urban blocks data and combine them to a single GeoDataFrame.

sample = geopandas.read_parquet("../data/sample.parquet")

all_poly = []
for i, row in sample.iterrows():
    fua = geopandas.read_parquet(f"../data/{int(row.eFUA_ID)}/polygons/")
    fua["continent"] = row.continent
    fua["country"] = row.Cntry_name
    fua["name"] = row.eFUA_name
    fua.crs = None
    all_poly.append(fua)
all_poly_data = pandas.concat(all_poly).reset_index(drop=True)

Plot settings and subfolders

mydirs = [
    "../plots/peaks/metrics/comparison",
    "../plots/peaks/metrics/circular_compactness",
]

for mydir in mydirs:
    os.makedirs(mydir, exist_ok=True)

rcParams.update({"figure.autolayout": True})


Plot peak results for all 5 metrics

for city in all_poly_data.name.unique():
    # initiate plot
    fig, ax = plt.subplots(
        1, 5, figsize=(20, 4), sharex=True, sharey=True, layout="constrained"
    )

    # get continent color from dict
    continent = numpy.max(all_poly_data[all_poly_data["name"] == city].continent)
    continent_color = sns.color_palette(n_colors=6)[coldict[continent]]

    for i, option in enumerate(options):
        option_label = option.replace("_", " ")

        my_pdf = numpy.array(results[city][option]["pdf"])

        mylinspace = numpy.linspace(
            start=results[city][option]["linspace_params"]["start"],
            stop=results[city][option]["linspace_params"]["stop"],
            num=results[city][option]["linspace_params"]["num"],
        )

        # plot vlines at identified peak (local max) points
        ax[i].vlines(
            x=mylinspace[results[city][option]["peaks"]],
            ymin=0,
            ymax=1,
            colors="black",
            linestyles="dotted",
            alpha=1,
            linewidth=1,
            zorder = 2
        )

        # plot vlines at identified valley (local min) points
        ax[i].vlines(
            x=mylinspace[results[city][option]["valleys"]],
            ymin=0,
            ymax=1,
            colors="red",
            linestyles="dotted",
            alpha=1,
            linewidth=1,
            zorder = 3
        )

        # plot pdf
        ax[i].plot(mylinspace, my_pdf, color=continent_color, zorder=1)

        # add scatter points at pdf/vline intersections
        if len(my_pdf[results[city][option]["peaks"]])>0:
            ax[i].scatter(
                x=mylinspace[results[city][option]["peaks"]],
                y=my_pdf[results[city][option]["peaks"]],
                color="black",
                s=8,
                alpha=1,
                zorder=4,
                label = "Peak"
            )

        if len(my_pdf[results[city][option]["valleys"]])>0:
            ax[i].scatter(
                x=mylinspace[results[city][option]["valleys"]],
                y=my_pdf[results[city][option]["valleys"]],
                color="red",
                s=8,
                alpha=1,
                zorder=5,
                label = "Valley"
            )

        ax[i].set_title(f"Peak detection in KDE based on {option_label}")
        ax[i].set_xlabel(f"log({option_label}) per area unit")
        ax[i].set_ylabel("Probability density")
        ax[i].set_xlim([-20, 20])
        ax[i].set_ylim([0, 1])
        ax[i].legend()

    # Set the font size of the figure title
    plt.rc("figure", titlesize=24)
    plt.suptitle(city)

    fig.savefig(f"../plots/peaks/metrics/comparison/{city}.png", dpi=300)

    plt.close()

Plot only circular compactness index metric (the one we choose to work with from now on) for each city

for city in all_poly_data.name.unique():
    # initiate plot
    fig, ax = plt.subplots(1, 1, figsize=(4, 4), layout="constrained")

    # get continent color from dict
    continent = numpy.max(all_poly_data[all_poly_data["name"] == city].continent)
    continent_color = sns.color_palette(n_colors=6)[coldict[continent]]

    option = "circular_compactness_index"
    option_label = option.replace("_", " ")

    my_pdf = numpy.array(results[city][option]["pdf"])

    mylinspace = numpy.linspace(
        start=results[city][option]["linspace_params"]["start"],
        stop=results[city][option]["linspace_params"]["stop"],
        num=results[city][option]["linspace_params"]["num"],
    )

    # plot vlines at identified peak (local max) points
    ax.vlines(
        x=mylinspace[results[city][option]["peaks"]],
        ymin=0,
        ymax=1,
        colors="black",
        linestyles="dotted",
        alpha=1,
        linewidth=1,
        zorder = 2
    )

    # plot vlines at identified valley (local min) points
    ax.vlines(
        x=mylinspace[results[city][option]["valleys"]],
        ymin=0,
        ymax=1,
        colors="red",
        linestyles="dotted",
        alpha=1,
        linewidth=1,
        zorder = 3
    )

    # plot pdf
    ax.plot(mylinspace, my_pdf, color=continent_color, zorder=1)

    # add scatter points at pdf/vline intersections
    if len(my_pdf[results[city][option]["peaks"]]) > 0:
        ax.scatter(
            x=mylinspace[results[city][option]["peaks"]],
            y=my_pdf[results[city][option]["peaks"]],
            color="black",
            s=8,
            alpha=1,
            zorder=4,
            label = "Peak", 
        )

    if len(my_pdf[results[city][option]["valleys"]]) > 0:
        ax.scatter(
            x=mylinspace[results[city][option]["valleys"]],
            y=my_pdf[results[city][option]["valleys"]],
            color="red",
            s=8,
            alpha=1,
            zorder=5,
            label = "Valley"
        )

    ax.set_title(f"Peak detection in KDE based on {option_label}")
    ax.set_xlabel(f"log({option_label}) per area unit")
    ax.set_ylabel("Probability density")
    ax.set_xlim([-20, 20])
    ax.set_ylim([0, 1])
    ax.legend()
    # Set the font size of the figure title
    plt.rc("figure", titlesize=24)
    plt.suptitle(city)

    fig.savefig(f"../plots/peaks/metrics/circular_compactness/{city}.png", dpi=300)

    plt.close()
